+++ spec/cinder_common_spec.rb
@@ -72,7 +72,7 @@ describe 'openstack-block-storage::cinder-common' do
       end
 
       it 'has proper modes' do
-        expect(sprintf('%o', file.mode)).to eq '644'
+        expect(sprintf('%o', file.mode)).to eq '640'
       end
 
       context 'template keystone contents' do