+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file is used to list changes made in each version of the openstack-block-st
 
 ## 10.0.0
 * Upgrading to Juno
+* Sync conf files with Juno
 
 ## 9.4.1
 * Add support for LVMISCSIDriver driver using block devices with LVM