+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 openstack-block-storage Cookbook CHANGELOG
 ==============================
 This file is used to list changes made in each version of the openstack-block-storage cookbook.
+## 10.0.1
+# Update cinder.conf mode from 0644 to 0640
 
 ## 10.0.0
 * Upgrading to Juno